[IMP] account: Test taxes results in both python/js from a single python test suite
The tests are currently duplicated between JS/Python.
Instead of that, the python is triggering a tour to gather the results from the JS side,
then compare the results.

task_id: 3775744
