+++ addons/account/controllers/__init__.py
@@ -4,3 +4,4 @@
 from . import portal
 from . import terms
 from . import download_docs
+from . import tests_shared_js_python
