+++ addons/account/tests/__init__.py
@@ -26,6 +26,7 @@ from . import test_digest
 from . import test_fiscal_position
 from . import test_sequence_mixin
 from . import test_tax
+from . import test_taxes_computation
 from . import test_invoice_taxes
 from . import test_account_move_send
 from . import test_account_all_l10n
