+++ addons/account/views/tests_shared_js_python.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<odoo>
+    <template id="tests_shared_js_python">
+        <t t-call="web.frontend_layout">
+            <t t-set="no_header" t-value="True"/>
+            <t t-set="no_footer" t-value="True"/>
+            <owl-component name="account.tests_shared_js_python" t-att-props="json.dumps(props)"/>
+        </t>
+    </template>
+</odoo>
