+++ addons/account_tax_python/tests/test_tax.py
@@ -17,9 +17,12 @@ class TestTaxPython(TestTaxCommon):
         })
 
     def test_python_taxes_for_l10n_in(self):
-        tax1 = self.python_tax("result = max(quantity * price_unit * 0.21, quantity * 4.17)")
-        self._check_tax_results(
+        tests = []
+        formula = "result = max(quantity * price_unit * 0.21, quantity * 4.17)"
+        tax1 = self.python_tax(formula)
+        tests.append(self._prepare_taxes_computation_test(
             tax1,
+            130.0,
             {
                 'total_included': 157.3,
                 'total_excluded': 130.0,
@@ -27,12 +30,12 @@ class TestTaxPython(TestTaxCommon):
                     (130.0, 27.3),
                 ),
             },
-            130.0,
-        )
+        ))
 
         tax1.price_include = True
-        self._check_tax_results(
+        tests.append(self._prepare_taxes_computation_test(
             tax1,
+            130.0,
             {
                 'total_included': 130.0,
                 'total_excluded': 102.7,
@@ -40,27 +43,27 @@ class TestTaxPython(TestTaxCommon):
                     (102.7, 27.3),
                 ),
             },
-            130.0,
-        )
+        ))
 
-        tax1.python_applicable = "False"
-        self._check_tax_results(
-            tax1,
+        tax2 = self.python_tax(formula, python_applicable="False")
+        tests.append(self._prepare_taxes_computation_test(
+            tax2,
+            130.0,
             {
                 'total_included': 130.0,
                 'total_excluded': 130.0,
                 'tax_values_list': [],
             },
-            130.0,
-        )
+        ))
 
         product1 = self.env['product.product'].create({
             'name': "product1",
             'volume': 200.0,
         })
-        tax2 = self.python_tax("result = product['volume'] > 100 and 10 or 5")
-        self._check_tax_results(
-            tax2,
+        tax3 = self.python_tax("result = product['volume'] > 100 and 10 or 5")
+        tests.append(self._prepare_taxes_computation_test(
+            tax3,
+            100.0,
             {
                 'total_included': 110.0,
                 'total_excluded': 100.0,
@@ -68,16 +71,16 @@ class TestTaxPython(TestTaxCommon):
                     (100.0, 10.0),
                 ),
             },
-            100.0,
             {'product': product1},
-        )
+        ))
 
         product2 = self.env['product.product'].create({
             'name': "product1",
             'volume': 50.0,
         })
-        self._check_tax_results(
-            tax2,
+        tests.append(self._prepare_taxes_computation_test(
+            tax3,
+            100.0,
             {
                 'total_included': 105.0,
                 'total_excluded': 100.0,
@@ -85,6 +88,6 @@ class TestTaxPython(TestTaxCommon):
                     (100.0, 5.0),
                 ),
             },
-            100.0,
             {'product': product2},
-        )
+        ))
+        self._assert_tests(tests, mode='py')
