+++ addons/l10n_in/tests/__init__.py
@@ -1,1 +1,1 @@
-from . import test_invoice_hsn_summary
+from . import test_hsn_summary
